write out zero duration for interesting alerts
